Move query names and Providers to parent module.
